Fix up the homing missile texture and maths
The texture was upside dowm, so the homing missile had a 180 rotation
built it, and the homing code took this into account.

This didn't work well with firepoints, so now the texture has been
fixed, and the 180 rotation taken out, and the homing code fixed up.

Now the rotation in the fire points matches the game properly